# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Platform Gateway API
  version: 1.0.0
  description: |
    **Platform Gateway**は、エンタープライズ級のマルチテナント管理プラットフォームです。
    Keycloak認証、Stripe Connect決済処理、Cloudflare Edge技術を統合した包括的なAPIを提供します。
    
    ## 主な特徴
    - 🔐 **Keycloak OIDC/JWT認証**: エンタープライズ級認証システム
    - 💳 **Stripe Connect統合**: マルチテナント決済処理
    - 🏢 **マルチテナント対応**: 組織単位でのデータ分離
    - 🛡️ **ロールベースアクセス制御**: 細かい権限管理
    - 📊 **リアルタイム分析**: ダッシュボード・メトリクス
    - 🌐 **Cloudflare Edge**: 高速・グローバル配信
    
    ## 認証方式
    - **JWT Bearer Token**: `Authorization: Bearer <token>`
    - **デモモード**: 開発・テスト用の認証なしアクセス
    
    ## レート制限
    - **一般API**: 100リクエスト/分
    - **認証API**: 20リクエスト/分
    - **決済API**: 50リクエスト/分
  contact:
    name: Platform Gateway API Support
    url: https://github.com/rixy-ynxy/rd-ai-agent-generator
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://ac94bb42.platform-gateway.pages.dev/api
    description: 本番環境
  - url: http://localhost:3000/api
    description: 開発環境
security:
  - bearerAuth: []
  - demoMode: []
paths:
...
  /users:
    get:
      tags: [ユーザー管理]
      summary: ユーザー一覧取得
      description: |
        テナント内のユーザー一覧を取得します。
        adminまたはmanager権限が必要。
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
          description: 対象テナントID (super_adminのみ)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, tenant_owner, admin, manager, developer, support, user]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
      responses:
        200:
          description: ユーザー一覧取得成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'

    post:
      tags: [ユーザー管理]
      summary: ユーザー作成
      description: |
        新しいユーザーを作成し、テナントに招待します。
        adminまたはmanager権限が必要。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        201:
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags: [ユーザー管理]
      summary: ユーザー詳細取得
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ユーザー取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'

    put:
      tags: [ユーザー管理]
      summary: ユーザー情報更新
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        200:
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'

    delete:
      tags: [ユーザー管理]
      summary: ユーザー削除
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ユーザー削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string

  # =============================================================================
  # 決済処理エンドポイント
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token (Keycloak発行)
    
    demoMode:
      type: apiKey
      in: header
      name: X-Demo-Mode
      description: デモモード用認証 (開発・テスト環境のみ)
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          description: データ配列
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: 総件数
            limit:
              type: integer
              description: 1ページあたりの件数
            offset:
              type: integer
              description: オフセット
            has_more:
              type: boolean
              description: 次ページの有無

    # =============================================================================
    # 認証関連
    # =============================================================================
...
    User:
      type: object
      properties:
        id:
          type: string
          description: ユーザーID
          example: "user-123abc"
        email:
          type: string
          format: email
          description: メールアドレス
          example: "user@example.com"
        name:
          type: string
          description: 表示名
          example: "田中太郎"
        username:
          type: string
          description: ユーザー名
          example: "tanaka_taro"
        tenant_id:
          type: string
          description: 所属テナントID
          example: "tenant-acme-corp"
        roles:
          type: array
          items:
            type: string
            enum: [super_admin, tenant_owner, admin, manager, developer, support, user]
          description: ユーザーロール
          example: ["admin", "manager"]
        status:
          type: string
          enum: [active, inactive, pending]
          description: ユーザーステータス
          example: "active"
        profile:
          type: object
          properties:
            avatar_url:
              type: string
              format: uri
            department:
              type: string
            job_title:
              type: string
            phone:
              type: string
        keycloak_id:
          type: string
          description: KeycloakでのユーザーID
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # =============================================================================
    # テナント関連
    # =============================================================================
...
    CreateUserRequest:
      type: object
      required: [email, name, roles]
      properties:
        email:
          type: string
          format: email
          example: "newuser@example.com"
        name:
          type: string
          example: "新規ユーザー"
        username:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [admin, manager, developer, support, user]
          example: ["user"]
        profile:
          type: object
          properties:
            department:
              type: string
            job_title:
              type: string
            phone:
              type: string
        send_invitation:
          type: boolean
          default: true
          description: 招待メールを送信するか
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, pending]
        profile:
          type: object

    # =============================================================================
    # 決済関連
    # =============================================================================
...